:truck:Refactor: move files 사이드바 컨텍스트 파일, 유틸 폴더 이동
