+++ tradin/src/components/units/backtest/main/BackTest.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import * as S from "./BackTest.styles";
-import { useSidebar } from "../../../commons/sidebar/SidebarContext";
+import { useSidebar } from "../../../../contexts/SidebarContext";
 import Breadcrumb from "../../../commons/breadcrumb/BreadCrumb.container";
 import OptionsContainer from '../option/BackTestOption';
 import ResultSkeletonUI from '../result/ResultSkeletonUI';
