+++ tradin/src/components/units/main/Main.presenter.tsx
@@ -1,7 +1,7 @@
 import * as S from "./Main.styles";
 import { useRecoilState } from "recoil";
-import { currencyKRW } from "../../commons/util/atoms";
-import { useSidebar } from "../../commons/sidebar/SidebarContext";
+import { currencyKRW } from "../../../util/atoms";
+import { useSidebar } from "../../../contexts/SidebarContext";
 import { FaPlus } from "react-icons/fa";
 import Widget from "./widget/Widget";
 import WidgetSelector from "./widget/WidgetSelector";
